Moved fonts, cleaning up CRTVideo, made textBox
