+++ modules/CRTVideo/font5x6.h
@@ -1,0 +1,6 @@
+#ifndef FONT5X6_H
+#define FONT5X6_H
+
+extern const unsigned char font5x6[];
+
+#endif
