+++ modules/CRTVideo/src/CRTVideo.cpp
@@ -4,6 +4,7 @@
 // driver
 #include "CRTVideo.h"
 #include "GraphicsDefault.h"
+#include "font5x6.h"
 
 // normal_distribution
 #include <iostream>
@@ -27,99 +28,9 @@ CRTVideo::CRTVideo(void)
 {
 }
 
-void CRTVideo::init(void)
-{
-	calculateAsciiLUT();
-	localPrintf("\n");
-	int offset = ('A'-0x20) * 48;
-	for(int i = offset; i < offset + 200; i++)
-	{
-		localPrintf("0x%02X ", asciiLUT[i]);
-		if(((i+1) % 6) == 0) localPrintf("\n");
-	}
-	bspDACInit();
-	interlace(false);
-}
-
 void CRTVideo::interlace(bool var)
 {
 	bspDACInterlace(var);
-}
-
-void CRTVideo::writeChar(char c)
-{
-	if(c == '\n')
-	{
-		cursorY++;
-		if(cursorY >= TEXT_MAP_HEIGHT - 1)
-		{
-			shiftTextUp();
-			cursorY = TEXT_MAP_HEIGHT - 2;
-		}
-		//Disabled for rx on line feed only
-//		return;
-//	}
-//	else if(c == '\r')
-//	{
-		cursorX = 1;
-		return;
-	}
-	else if(c == 0x08)//backspace
-	{
-		cursorX--;
-		if(cursorX <= 0)
-		{
-			//Special case, no up sroll
-			if(cursorY <= 1)
-			{
-				cursorX = 1;
-				return;
-			}
-			else if(cursorY > 1)
-			{
-				cursorY--;
-				cursorX = TEXT_MAP_WIDTH - 2;
-			}
-		}
-		textMap[(TEXT_MAP_WIDTH * cursorY) + cursorX] = ' ';
-		return;
-	}
-	textMap[(TEXT_MAP_WIDTH * cursorY) + cursorX] = c;
-	cursorX++;
-	if(cursorX >= TEXT_MAP_WIDTH - 1)
-	{
-		cursorX = 1;
-		cursorY++;
-		if(cursorY >= TEXT_MAP_HEIGHT - 1)
-		{
-			shiftTextUp();
-			cursorY = TEXT_MAP_HEIGHT - 2;
-		}
-	}
-}
-
-//Optional/example call to write text to screen.
-//Alternatly, take your own frame and use console()
-void CRTVideo::drawFrame(void)
-{
-	bspIOPinWrite(D31, 0);
-	
-	uint8_t textX = 1;
-	uint8_t textY = 1;
-
-	uint8_t * nextFrame = 0;
-	if(bspDACGetBufferBlank(&nextFrame, 0x00)) //address of pointer
-	{
-		for(textY = 0; textY < TEXT_MAP_HEIGHT; textY++)
-		{
-			for(textX = 0; textX < TEXT_MAP_WIDTH; textX++)
-			{
-				writeAscii5x7(nextFrame, textX * 6, textY * 8, textMap[(textY * TEXT_MAP_WIDTH) + textX]);
-			}
-		}
-		bspDACSwapBuffers();
-	}
-	bspIOPinWrite(D31, 1);
 }
 
 bool CRTVideo::getBlank(uint8_t ** output, uint8_t fill)
@@ -136,20 +47,6 @@ void CRTVideo::swap(void)
 	bspDACSwapBuffers();
 }
 
-void CRTVideo::console(uint8_t * dst)
-{
-	uint8_t textX = 1;
-	uint8_t textY = 1;
-
-	for(textY = 0; textY < TEXT_MAP_HEIGHT; textY++)
-	{
-		for(textX = 0; textX < TEXT_MAP_WIDTH; textX++)
-		{
-			writeAscii5x7(dst, textX * 6, textY * 8, textMap[(textY * TEXT_MAP_WIDTH) + textX]);
-		}
-	}
-}
-
 bool CRTVideo::pixel(uint8_t * dst, uint8_t x, uint8_t y, uint8_t value)
 {
 	if((x >= PIXEL_WIDTH)||(y >= PIXEL_HEIGHT))
@@ -163,13 +60,6 @@ bool CRTVideo::pixel(uint8_t * dst, uint8_t x, uint8_t y, uint8_t value)
 
 bool CRTVideo::line(uint8_t * dst, uint8_t x1, uint8_t y1, uint8_t x2, uint8_t y2, uint8_t value)
 {
-	//if((x1 > x2)||(y1 > y2))
-	//{
-	//	return false;
-	//}
-	//draw vertical segments on each x
-	//Swap x1, x2 if needed
-	
 	if(x1 < x2)
 	{
 		for(int i = x1; i < x2; i++)
@@ -218,42 +108,6 @@ bool CRTVideo::line(uint8_t * dst, uint8_t x1, uint8_t y1, uint8_t x2, uint8_t y
 			}
 		}
 	}
-	
-	
-	// eh?
-	
-	//if(x1 > x2)
-	//{
-	//	int tmp = x1;
-	//	x1 = x2;
-	//	x2 = tmp;
-	//	tmp = y1;
-	//	y1 = y2;
-	//	y2 = tmp;
-	//}
-	//for( int i = 0; i <= (x2 - x1); i++ )
-	//{
-	//	int16_t start;
-	//	int16_t end;
-	//	if(y2 > y1)
-	//	{
-	//		start = y1 + (((float)i/(float)(x2 - x1)) * (float)(y2 - y1));
-	//		end = y1 + (((float)(i + 1)/(float)(x2 - x1)) * (float)(y2 - y1));
-	//		for( int j = start; j <= end; j++ )
-	//		{
-	//			pixel(dst, x1 + i, j, 0xFF);
-	//		}
-	//	}
-	//	else
-	//	{
-	//		start = y2 + (((float)i/(float)(x2 - x1)) * (float)(y1 - y2));
-	//		end = y2 + (((float)(i + 1)/(float)(x2 - x1)) * (float)(y1 - y2));
-	//		for( int j = start; j <= end; j++ )
-	//		{
-	//			pixel(dst, x1 + i, j, 0xFF);
-	//		}
-	//	}
-	//}
 	return true;
 }
 
@@ -328,31 +182,38 @@ bool CRTVideo::drawBitmap(uint8_t * dst, virtual_bitmap_type_t * pBitmap, int16_
 	return true;
 }
 
-void CRTVideo::shiftTextUp(void)
-{
-	int i;
-	for(i = 1; i < TEXT_MAP_HEIGHT - 1; i++)
-	{
-		memcpy(&textMap[(i*TEXT_MAP_WIDTH)], &textMap[((i+1)*TEXT_MAP_WIDTH)], TEXT_MAP_WIDTH);
-	}
-	memset(&textMap[((i-1)*TEXT_MAP_WIDTH)], ' ', TEXT_MAP_WIDTH);
-}
-
-void CRTVideo::writeAscii5x7(uint8_t * dst, uint8_t dstX, uint8_t dstY, char c)
+void CRTVideo::drawAscii(uint8_t * dst, uint8_t dstX, uint8_t dstY, char c, font_t * font)
 {
 	int line;
-	if((c >= 0x20)&&(c < 0x7F))
+	if((c >= font->startC)&&(c < (0x20 + font->lenC)))
 	{
 		//printable
 	}
 	else
 	{
-		c = 0x20; //default print space
+		c = font->startC; //default print char
 	}
 	
-	for(line = 0; line < 8; line++)
+	int pxPerC = font->charWPx * font->charHPx;
+	for(line = 0; line < font->charHPx; line++)
 	{
-		//8 * 6 = 48
-		memcpy(&dst[((dstY + line) * 192) + dstX], &asciiLUT[((c-0x20) * 48) + (line * 6)], 6);
+        int writeIndex = ((dstY + line) * PIXEL_WIDTH) + dstX;
+        if(writeIndex < PIXEL_TOT - font->charWPx)
+        {
+            //weak check, at least it's on the array data
+            memcpy(&dst[writeIndex], &(font->data[((c-font->startC) * pxPerC) + (line * font->charWPx)]), font->charWPx);
+        }
+	}
+}
+
+void CRTVideo::drawAscii(uint8_t * dst, uint8_t xOrigin, uint8_t yOrigin, const char * str, uint8_t xChar, uint8_t yChar, font_t * font)
+{
+	int i = 0;
+	yOrigin = yOrigin + (yChar * font->charHPx);
+	while(*str != 0)
+	{
+		drawAscii(dst, xOrigin + ((i + xChar) * font->charWPx), yOrigin, *str, font);
+		i++;
+		str++;
 	}
 }
