+++ modules/CRTVideo/src/GraphicsDefault.h
@@ -4,11 +4,6 @@
 extern "C" {
 #endif
 
-#define ASCII_BLACK_LEVEL 98
-
-extern uint8_t asciiLUT[];
-
-void calculateAsciiLUT(void);
 
 #ifdef __cplusplus
 }
